Add separate tests for existing project name
Concretely, the first test should ensure that an exception is thrown
when the new application has the same name and project with
the existent one. The second test should make sure that everything
is ok when the new application has the same name but a different project
from the existent one.